set sample time as int

+++ dev/test_chains.py
@@ -16,7 +16,7 @@ def transform(df: pd.DataFrame) -> pd.DataFrame:
     This function is passed to the load_data_to_mariadb function
     """
     # Convert milliseconds to seconds *in place*
-    df['time'] = df['time'] / 1000  
+    df['time'] = (df['time'] / 1000).astype(int)  
     # Create datetime and date objects using the correct column
     df['quote_time'] = pd.to_datetime(df['time'], unit='s')
     df['quote_date'] = df['quote_time'].dt.date
